+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.0] - 2026-02-22
 - fix #9, fix sign bit for 12, 14, 16 bit reads.
+- add writeConfig() of defaults in begin()
 - minor edits
 
 ----